Last-ditch exception handling.
Deal with uncaught errors at the last moment, in lineReceived(). Log the traceback serverside. If the player's debug attr is True, present it to them as well; otherwise apologize vaguely.

This is intended for code errors, not exceptional cases or user mistakes. If our code were perfectly written, this block would never be run.