Add a Python REPL. Do not use without reading this commit message in full.
So far, since there's no support for untrusted user code, the only REPL that makes sense is a trusted REPL. That is, interactive access to *the very Python process* that is the game server itself. This is now available via "sudo python."

If you haven't just soiled your pants in terror, you haven't understood properly.

Direct interactive access means that it's practical to very carefully administer a live game: you can move objects around, edit attributes freely, adjust permissions, and so on, all with SYSTEM access, by means of retrieving the objects from the database and futzing with them. It also means it's practical to debug the server software, which is going to be lovely while we're in development.

It also means that "del muss.db._objects" would run without complaint.

So would "os.system('rm -rf $HOME')".

Treat this function with extreme respect. Even if you never use it, its mere existence is a very real privilege-escalation threat if this project ever gets big enough to attract the people who are into that sort of thing.