Replace our LineReceiver with a TelnetProtocol.
Our dataReceived() method buffers input, splits into lines, and calls lineReceived() on each, just as a LineReceiver would, and our sendLine() method tacks on a delimiter. The new Protocol is a drop-in replacement for the old one.

The advantage is that it's also capable of doing telnet settings negotiation for screen width and so on.