Remove the name field from OneOf's constructor.
Calling setName() is more idiomatic in pyparsing, so the constructor field is
extraneous (even though I think it's a better interface). Some sort of name is
more or less required, though, so we provide a fairly useless one for OneOf and
SomeOf, with the expectation that it'll get overridden in all but test code.
(Which is why I like the constructor field better. Oh well.)