Use MUSSTestCase more consistently.
Every test case that uses the DB at all should be a subclass of
common_tools.MUSSTestCase, not of unittest.TestCase. That way, the DB is mocked
out consistently, and there's no risk of tests passing or failing depending on
whether a muss.db file is present or what's in it.